+++ src/runner/mod.rs
@@ -3,9 +3,11 @@
 mod engine;
 mod errors;
 mod execution;
+mod interrupt;
 mod memory;
 mod module_loader;
 mod promise_rejections;
 mod transpile;
 
 pub use engine::Engine;
+pub use interrupt::InterruptHandle;